+++ src/main/ipcs/server/rcon/sendRCONCommand.ts
@@ -1,22 +1,16 @@
-import { ipcMain, ipcRenderer } from 'electron';
+import { ipcMain } from 'electron';
 import Channels from '../../channels';
-import trimWorldSettingsString from '../../../../utils/trimWorldSettingsString';
-import axios from 'axios';
-import getWorldSettingsByServerId from '../../../services/worldSettings/getWorldSettingsByServerId';
 import sendCommand from '../../../utils/rcon/sendCommand';
+import getAdminConnectionConfig, {
+  getRconOptions,
+} from '../../../services/admin/getAdminHost';
 
 ipcMain.handle(
   Channels.sendRCONCommand,
   async (event, serverId: string, command: string) => {
     try {
-      const worldSettings = await getWorldSettingsByServerId(serverId);
-
-      const serverOptions = {
-        ipAddress: '127.0.0.1',
-        port: worldSettings.RCONPort,
-        password: trimWorldSettingsString(worldSettings.AdminPassword),
-      };
-      const response = await sendCommand(serverOptions, command);
+      const connection = await getAdminConnectionConfig(serverId);
+      const response = await sendCommand(getRconOptions(connection), command);
       return response;
     } catch (e) {}
   },